front
Izmjene

+++ FRONTEND/src/service/PrimkaServisaService.js
@@ -1,62 +1,56 @@
+// PrimkaServisaService.js
 import { HttpService } from './HttpService';
 
-const ime= '/PrimkaServisa';
+const ime = '/PrimkaServisa';
 
 async function get() {
-    return await HttpService.get(ime)
-        .then((response) => {
-            return { greska: false, poruka: response.data };
-        })
-        .catch((error) => {
-            return { greska: true, poruka: error };
-        });
+    try {
+        const odgovor = await HttpService.get(ime);
+        return { greska: false, poruka: odgovor.data };
+    } catch (e) {
+        return { greska: true, poruka: e };
+    }
 }
 
 async function post(primkaServisa) {
-    return await HttpService.post(ime, primkaServisa)
-        .then((response) => {
-            return { greska: false, poruka: response.data };
-        })
-        .catch((error) => {
-            return { greska: true, poruka: error };
-        });
+    try {
+        const odgovor = await HttpService.post(ime, primkaServisa);
+        return { greska: false, poruka: odgovor.data };
+    } catch (e) {
+        return { greska: true, poruka: e };
+    }
 }
 
 async function put(id, primkaServisa) {
-    return await HttpService.put(ime + '/'+id,primkaServisa)
-    .then((odgovor)=>{
-        //console.table(odgovor.data);
-        return {greska: false, poruka: odgovor.data};
-    })
-    .catch((e)=>{
-        //console.log(e);
-        return {greska: true, poruka: e};
-    })
+    try {
+        const odgovor = await HttpService.put(ime + '/' + id, primkaServisa);
+        return { greska: false, poruka: odgovor.data };
+    } catch (e) {
+        return { greska: true, poruka: e };
+    }
 }
 
-async function _delete(idPrimkaServisa){
-    return await HttpService.delete(ime + '/'+ idPrimkaServisa)
-    .then((odgovor)=>{
-        //console.table(odgovor.data);
-        return {greska: false, poruka: odgovor.data.poruka};
-    })
-    .catch((e)=>{
-        //console.log(e);
-        return {greska: true, poruka: e};
-    })
+async function _delete(idPrimkaServisa) {
+    try {
+        const povezaniServisniNalozi = await HttpService.get('/ServisniNalozi/?primkaServisaId=' + idPrimkaServisa);
+        if (povezaniServisniNalozi.data.length > 0) {
+            return { greska: true, poruka: 'Ne možete obrisati ovu primku servisa jer je povezana s jednim ili više servisnih naloga.' };
+        }
+        const odgovor = await HttpService.delete(ime + '/' + idPrimkaServisa);
+        return { greska: false, poruka: odgovor.data.poruka };
+    } catch (e) {
+        return { greska: true, poruka: e };
+    }
 }
 
-async function getById(id){
-    return await HttpService.get(ime+'/'+id)
-    .then((o)=>{
-        return {greska: false, poruka: o.data}
-    })
-    .catch((e)=>{
-        return {greska: true, poruka: e}
-    });
+async function getById(id) {
+    try {
+        const odgovor = await HttpService.get(ime + '/' + id);
+        return { greska: false, poruka: odgovor.data };
+    } catch (e) {
+        return { greska: true, poruka: e };
+    }
 }
-
-
 
 export default {
     get,
